# openapi.yaml
openapi: 3.0.3
info:
  title: Watch Merchant API
  description: |
    Demo watch merchant API with static product catalog for AP2 integration.

    This API provides endpoints for browsing luxury watches, managing shopping carts,
    calculating pricing, checking inventory, and placing orders.
  version: 1.0.0
  contact:
    name: API Support
    email: support@watchmerchant.demo
servers:
  - url: http://localhost:8765/api/v1
    description: Local development server
  - url: https://api.watchmerchant.demo/v1
    description: Production server
...
paths:
...
  /cart:
    post:
      tags: [Cart]
      summary: Create or update cart
      description: Create a new shopping cart or update an existing one
      operationId: watches_createCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cartId:
                  type: string
                  description: Existing cart ID (optional)
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      sku:
                        type: string
                      quantity:
                        type: integer
                        minimum: 1
              required: [items]
      responses:
        '201':
          description: Cart created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cart/{cartId}:
    get:
      tags: [Cart]
      summary: Get cart
      description: Retrieve shopping cart contents
      operationId: watches_getCart
      parameters:
        - name: cartId
          in: path
          required: true
          description: Cart ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Cart]
      summary: Delete cart
      description: Remove shopping cart
      operationId: watches_deleteCart
      parameters:
        - name: cartId
          in: path
          required: true
          description: Cart ID
          schema:
            type: string
      responses:
        '200':
          description: Cart deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /cart/{cartId}/items/{sku}:
    put:
      tags: [Cart]
      summary: Update cart item
      description: Update quantity of an item in cart (set to 0 to remove)
      operationId: watches_updateCartItem
      parameters:
        - name: cartId
          in: path
          required: true
          description: Cart ID
          schema:
            type: string
        - name: sku
          in: path
          required: true
          description: Product SKU
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 0
              required: [quantity]
      responses:
        '200':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /cart/{cartId}/create-mandate:
    post:
      tags: [Cart]
      summary: Create payment mandate
      description: Generate EIP-712 typed data for MetaMask payment signing
      operationId: watches_createMandate
      parameters:
        - name: cartId
          in: path
          required: true
          description: Cart ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  description: User's wallet address
                network:
                  type: string
                  description: Network ID (e.g., "sepolia", "bnb-mainnet")
              required: [walletAddress, network]
      responses:
        '200':
          description: Payment mandate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cart/{cartId}/verify-payment:
    post:
      tags: [Cart]
      summary: Verify payment
      description: Verify MetaMask signature and create order
      operationId: watches_verifyPayment
      parameters:
        - name: cartId
          in: path
          required: true
          description: Cart ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                signature:
                  type: string
                  description: MetaMask signature
                mandateHash:
                  type: string
                  description: Mandate hash from create-mandate
              required: [signature, mandateHash]
      responses:
        '200':
          description: Payment verified and order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /inventory/check:
    post:
      tags: [Inventory]
      summary: Check inventory for multiple SKUs
      description: Bulk check stock availability
      operationId: watches_checkInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skus:
                  type: array
                  items:
                    type: string
              required: [skus]
      responses:
        '200':
          description: Inventory status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryStatus'
  /inventory/{sku}:
    get:
      tags: [Inventory]
      summary: Check inventory for single SKU
      description: Check stock availability for a specific product
      operationId: watches_checkInventorySingle
      parameters:
        - name: sku
          in: path
          required: true
          description: Product SKU
          schema:
            type: string
      responses:
        '200':
          description: Inventory status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/InventoryStatus'
...
  /orders:
    post:
      tags: [Orders]
      summary: Create order
      description: Place a new order from a cart
      operationId: watches_createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cartId:
                  type: string
                customer:
                  $ref: '#/components/schemas/CustomerInfo'
                shipping:
                  $ref: '#/components/schemas/ShippingInfo'
                payment:
                  type: object
                  properties:
                    method:
                      type: string
                    token:
                      type: string
                  required: [method]
                couponCode:
                  type: string
              required: [cartId, customer, shipping, payment]
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order
      description: Retrieve order details
      operationId: watches_getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
  /orders/{orderId}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      description: Update the status of an order
      operationId: watches_updateOrderStatus
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending_payment, processing, shipped, delivered, cancelled]
              required: [status]
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
  /orders/{orderId}/returns:
    post:
      tags: [Orders]
      summary: Create return request
      description: Request a return for order items
      operationId: watches_createReturn
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      sku:
                        type: string
                      quantity:
                        type: integer
                      reason:
                        type: string
                    required: [sku, quantity, reason]
                customerNotes:
                  type: string
              required: [items]
      responses:
        '201':
          description: Return created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  schemas:
...
    PricingBreakdown:
      type: object
      properties:
        subtotal:
          type: number
        tax:
          type: number
        shipping:
          type: number
        discount:
          type: number
        total:
          type: number
        shippingOptions:
          type: array
          items:
            $ref: '#/components/schemas/ShippingRate'
    ShippingRate:
      type: object
      properties:
        carrier:
          type: string
        service:
          type: string
        cost:
          type: number
        estimatedDays:
          type: string
    InventoryStatus:
      type: object
      properties:
        sku:
          type: string
        available:
          type: boolean
        quantity:
          type: integer
        leadTime:
          type: string
    CustomerInfo:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required: [email, firstName, lastName]
    ShippingInfo:
      type: object
      properties:
        address:
          type: object
          properties:
            line1:
              type: string
            line2:
              type: string
            city:
              type: string
            state:
              type: string
            postalCode:
              type: string
            country:
              type: string
          required: [line1, city, state, postalCode, country]
        method:
          type: string
      required: [address, method]
    Order:
      type: object
      properties:
        orderId:
          type: string
        cartId:
          type: string
        status:
          type: string
          enum: [pending_payment, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            type: object
        customer:
          $ref: '#/components/schemas/CustomerInfo'
        shipping:
          $ref: '#/components/schemas/ShippingInfo'
        pricing:
          $ref: '#/components/schemas/PricingBreakdown'
        payment:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object